changed login folder structure
