+++ src/frontend/components/login-card/index.ts
@@ -1,0 +1,2 @@
+export * from './login-card';
+export { default } from './login-card';
